GITIGNORE - remove and ignore with-redux-app/ dir
I used a boilerplate next.js install to help me figure out some
configs with redux. I don't need it anymore, so I'm just cleaning
house a bit. I'm not sure exactly which one it was, but
something similar to:

$ npx create-next-app redux-next-app

This not the exact source I used, but again, it was something
similar to https://vpilip.com/how-setup-nextjs-with-redux-and-redux-debuger/